backend/requirements.txt
# 智能旧衣回收箱系统 - Python后端依赖

# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
aiomysql==0.2.0
pymysql==1.1.0
redis==5.0.1

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1

# 工具库
python-dateutil==2.8.2
loguru==0.7.2
python-dotenv==1.0.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3

